+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Add GetAllCmdArgs to StageContext
+
 ## [0.5.3] - 2023-09-14
 
 - Fix: help information is not printed for whenNot